# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/review-cycles": {
      "get": {
        "operationId": "Reviews_cycles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      },
      "post": {
        "operationId": "Reviews_createCycle",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "okrCycleId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "periodEnd": {
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "type": "string"
                  },
                  "periodStart": {
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "type": "string"
                  },
                  "population": {
                    "additionalProperties": false,
                    "default": {},
                    "properties": {
                      "excludeHiredAfter": {
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                        "type": "string"
                      },
                      "excludePersonIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "orgUnitIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "personIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      }
                    },
                    "type": "object"
                  },
                  "templateId": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "templateId",
                  "name",
                  "periodStart",
                  "periodEnd"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/review-cycles/{id}": {
      "get": {
        "operationId": "Reviews_cycle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      },
      "patch": {
        "operationId": "Reviews_updateCycle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "managerDueAt": {
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "okrCycleId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "population": {
                    "additionalProperties": false,
                    "properties": {
                      "excludeHiredAfter": {
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                        "type": "string"
                      },
                      "excludePersonIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "orgUnitIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "personIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      }
                    },
                    "type": "object"
                  },
                  "selfDueAt": {
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/review-cycles/{id}/close": {
      "post": {
        "operationId": "Reviews_close",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/review-cycles/{id}/launch": {
      "post": {
        "operationId": "Reviews_launch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "launchDate": {
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/review-cycles/{id}/population": {
      "get": {
        "operationId": "Reviews_population",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Anteprima popolazione: inclusi, esclusi e persone senza manager",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/review-cycles/{id}/progress": {
      "get": {
        "operationId": "Reviews_progress",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/review-cycles/{id}/remind": {
      "post": {
        "operationId": "Reviews_remind",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/review-templates": {
      "get": {
        "operationId": "Reviews_templates",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      },
      "post": {
        "operationId": "Reviews_createTemplate",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "description": {
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "includeObjectives": {
                    "default": true,
                    "type": "boolean"
                  },
                  "managerDueDays": {
                    "default": 21,
                    "maximum": 365,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "managerFormKey": {
                    "pattern": "^[a-z][a-z0-9_]{0,60}$",
                    "type": "string"
                  },
                  "managerSeesSelf": {
                    "default": "after_submit",
                    "enum": [
                      "immediately",
                      "after_submit",
                      "never"
                    ],
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "overallRatingField": {
                    "maxLength": 60,
                    "nullable": true,
                    "type": "string"
                  },
                  "ratingScale": {
                    "additionalProperties": false,
                    "properties": {
                      "labels": {
                        "additionalProperties": {
                          "type": "string"
                        },
                        "default": {},
                        "type": "object"
                      },
                      "max": {
                        "type": "integer"
                      },
                      "min": {
                        "type": "integer"
                      }
                    },
                    "required": [
                      "min",
                      "max"
                    ],
                    "type": "object"
                  },
                  "requireSignature": {
                    "default": true,
                    "type": "boolean"
                  },
                  "selfDueDays": {
                    "default": 14,
                    "maximum": 365,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "selfFormKey": {
                    "nullable": true,
                    "pattern": "^[a-z][a-z0-9_]{0,60}$",
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "managerFormKey"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Template di review: form self/manager (chiavi del form engine), fasi, scadenze, regole",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/review-templates/{id}": {
      "patch": {
        "operationId": "Reviews_updateTemplate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "archived": {
                    "type": "boolean"
                  },
                  "description": {
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "includeObjectives": {
                    "default": true,
                    "type": "boolean"
                  },
                  "managerDueDays": {
                    "default": 21,
                    "maximum": 365,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "managerFormKey": {
                    "pattern": "^[a-z][a-z0-9_]{0,60}$",
                    "type": "string"
                  },
                  "managerSeesSelf": {
                    "default": "after_submit",
                    "enum": [
                      "immediately",
                      "after_submit",
                      "never"
                    ],
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "overallRatingField": {
                    "maxLength": 60,
                    "nullable": true,
                    "type": "string"
                  },
                  "ratingScale": {
                    "additionalProperties": false,
                    "properties": {
                      "labels": {
                        "additionalProperties": {
                          "type": "string"
                        },
                        "default": {},
                        "type": "object"
                      },
                      "max": {
                        "type": "integer"
                      },
                      "min": {
                        "type": "integer"
                      }
                    },
                    "required": [
                      "min",
                      "max"
                    ],
                    "type": "object"
                  },
                  "requireSignature": {
                    "default": true,
                    "type": "boolean"
                  },
                  "selfDueDays": {
                    "default": 14,
                    "maximum": 365,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "selfFormKey": {
                    "nullable": true,
                    "pattern": "^[a-z][a-z0-9_]{0,60}$",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews": {
      "get": {
        "operationId": "Reviews_list",
        "parameters": [
          {
            "in": "query",
            "name": "box",
            "required": false,
            "schema": {
              "default": "mine",
              "enum": [
                "mine",
                "team",
                "all"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cycleId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{id}": {
      "get": {
        "operationId": "Reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{id}/context": {
      "get": {
        "operationId": "Reviews_context",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pannello di contesto: obiettivi, feedback condivisi, riconoscimenti, review precedenti, 1:1",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{id}/conversation": {
      "post": {
        "operationId": "Reviews_conversation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "at": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "notes": {
                    "maxLength": 4000,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{id}/rating-override": {
      "post": {
        "operationId": "Reviews_override",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "note": {
                    "maxLength": 1000,
                    "minLength": 3,
                    "type": "string"
                  },
                  "rating": {
                    "type": "integer"
                  }
                },
                "required": [
                  "rating",
                  "note"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{id}/reopen": {
      "post": {
        "operationId": "Reviews_reopen",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "stage": {
                    "enum": [
                      "self",
                      "manager"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "stage"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{id}/share": {
      "post": {
        "operationId": "Reviews_share",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "WorkingBetter API",
    "description": "API REST unica per web, mobile e connettori (ADR-0005). Errori in formato RFC 9457 (application/problem+json). I body e le query sono descritti dagli stessi schemi Zod usati per la validazione (ADR-0009).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "URI del tipo di errore (about:blank per gli errori generici)"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "description": "Identificativo della richiesta (x-request-id)"
          },
          "code": {
            "type": "string",
            "description": "Codice applicativo stabile (ErrorCodes in @wb/shared)"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "path",
                "message"
              ],
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}